+++ internal/db/coin_transfer_access.go
@@ -2,7 +2,6 @@ package db
 
 import (
 	"context"
-	"log"
 
 	"github.com/google/uuid"
 	"github.com/jackc/pgx/v5"
@@ -12,9 +11,9 @@ import (
 
 func GetUserTransactionHistory(ctx context.Context, pool *pgxpool.Pool, username string) (recieved []model.CoinTransfer, sent []model.CoinTransfer, err error) {
 	query := `
-		SELECT id, sender_id, receiver_id, amount
+		SELECT id, sender_login, receiver_login, amount
 		FROM coin_transfer
-		WHERE receiver_id = $1
+		WHERE receiver_login = $1
 	`
 
 	rows, err := pool.Query(ctx, query, username)
@@ -32,9 +31,9 @@ func GetUserTransactionHistory(ctx context.Context, pool *pgxpool.Pool, username
 	}
 
 	query = `
-		SELECT id, sender_id, receiver_id, amount
+		SELECT id, sender_login, receiver_login, amount
 		FROM coin_transfer
-		WHERE sender_id = $1
+		WHERE sender_login = $1
 	`
 
 	rows, err = pool.Query(ctx, query, username)
@@ -56,13 +55,12 @@ func GetUserTransactionHistory(ctx context.Context, pool *pgxpool.Pool, username
 
 func InsertCoinTransferTx(ctx context.Context, tx pgx.Tx, sender string, reciever string, amount int64) error {
 	query := `
-		INSERT INTO coin_transfer (id, sender_id, receiver_id, amount) 
+		INSERT INTO coin_transfer (id, sender_login, receiver_login, amount) 
 		VALUES ($1, $2, $3, $4)
 	`
 
 	_, err := tx.Exec(ctx, query, uuid.New(), sender, reciever, amount)
 	if err != nil {
-		log.Printf("Failed to insert coin transfer, sender: %v, reciever: %v error: %v", sender, reciever, err)
 		return err
 	}
